Move ControlStyles type to control-styles
