+++ editor/src/components/inspector/common/property-path-hooks.ts
@@ -16,9 +16,10 @@ import {
 } from '../../../components/editor/store/store-hook'
 import type {
   ControlStatus,
-  ControlStyles,
   PropertyStatus,
 } from '../../../components/inspector/common/control-status'
+import type { ControlStyles } from '../../../components/inspector/common/control-styles'
+
 import {
   calculateMultiPropertyStatusForSelection,
   calculateMultiStringPropertyStatusForSelection,
