+++ editor/src/components/inspector/controls/background-solid-or-gradient-thumbnail-control.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import type { UseSubmitValueFactory } from '../common/property-path-hooks'
 import { isRight } from '../../../core/shared/either'
-import type { ControlStatus, ControlStyles } from '../common/control-status'
+import type { ControlStyles } from '../common/control-styles'
+import type { ControlStatus } from '../common/control-status'
 import type {
   CSSBackground,
   CSSBackgroundLayer,
