+++ editor/src/components/inspector/controls/color-control.tsx
@@ -4,7 +4,8 @@ import { ColorPicker } from './color-picker'
 import type { CSSColor } from '../common/css-utils'
 import { parseColor, cssColorToChromaColorOrDefault, printColor } from '../common/css-utils'
 import { StringControl } from './string-control'
-import type { ControlStatus, ControlStyles } from '../common/control-status'
+import type { ControlStyles } from '../common/control-styles'
+import type { ControlStatus } from '../common/control-status'
 import { useColorTheme, UtopiaTheme } from '../../../uuiui'
 import Utils from '../../../utils/utils'
