+++ editor/src/components/inspector/controls/conditional-override-control.tsx
@@ -3,7 +3,8 @@ import type { ConditionValue } from '../../../core/shared/element-template'
 import type { ButtonProps } from '../../../uuiui'
 import { FlexRow, Icons, Tooltip } from '../../../uuiui'
 import { SquareButton } from '../../titlebar/buttons'
-import type { ControlStatus, ControlStyles } from '../common/control-status'
+import type { ControlStyles } from '../common/control-styles'
+import type { ControlStatus } from '../common/control-status'
 import { UIGridRow } from '../widgets/ui-grid-row'
 import type { OptionChainOption } from './option-chain-control'
 import { OptionChainControl } from './option-chain-control'
