+++ editor/src/components/inspector/controls/control.ts
@@ -2,7 +2,8 @@ import type { OptionChainOption } from './option-chain-control'
 import type { DEPRECATEDOptionControlOptions } from './option-control'
 import type { DEPRECATEDSelectControlOptions, SelectOption } from './select-control'
 import type { DEPRECATEDSliderControlOptions } from './slider-control'
-import type { ControlStatus, ControlStyles } from '../common/control-status'
+import type { ControlStyles } from '../common/control-styles'
+import type { ControlStatus } from '../common/control-status'
 import type { UnknownOrEmptyInput, EmptyInputValue } from '../common/css-utils'
 import type { StringControlOptions } from './string-control'
