+++ editor/src/components/inspector/controls/unknown-array-item.tsx
@@ -4,7 +4,8 @@ import { isCSSUnknownArrayItem, isEmptyInputValue } from '../common/css-utils'
 import { CSSArrayItem, getIndexedSpliceArrayItem } from '../common/inspector-utils'
 import type { UseSubmitValueFactory } from '../common/property-path-hooks'
 import { stopPropagation } from '../common/inspector-utils'
-import type { ControlStatus, ControlStyles } from '../common/control-status'
+import type { ControlStyles } from '../common/control-styles'
+import type { ControlStatus } from '../common/control-status'
 import { PropertyRow } from '../widgets/property-row'
 import { StringControl } from './string-control'
 import { CheckboxInput, FlexRow, Tooltip } from '../../../uuiui'
