+++ editor/src/components/inspector/sections/layout-section/flex-container-subsection/flex-container-controls.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import type { FlexAlignment, FlexJustifyContent } from 'utopia-api/core'
 import { FlexWrap } from 'utopia-api/core'
-import type { ControlStatus, ControlStyles } from '../../../common/control-status'
+import type { ControlStatus } from '../../../common/control-status'
+import type { ControlStyles } from '../../../common/control-styles'
 import { InspectorContextMenuWrapper } from '../../../../context-menu-wrapper'
 import type { OptionChainOption } from '../../../controls/option-chain-control'
 import { OptionChainControl } from '../../../controls/option-chain-control'
