+++ editor/src/components/inspector/sections/layout-section/layout-system-subsection/layout-system-controls.tsx
@@ -25,11 +25,6 @@ import {
   useSelectorWithCallback,
 } from '../../../../editor/store/store-hook'
 import { optionalAddOnUnsetValues } from '../../../common/context-menu-items'
-import {
-  ControlStatus,
-  ControlStyles,
-  getControlStatusFromPropertyStatus,
-} from '../../../common/control-status'
 import { getControlStyles } from '../../../common/control-styles'
 import type { CanvasControlWithProps } from '../../../common/inspector-atoms'
 import { useControlModeWithCycle } from '../../../common/inspector-utils'
