+++ editor/src/components/inspector/sections/style-section/text-subsection/text-shadow-subsection.tsx
@@ -3,7 +3,8 @@ import type { SpringValue } from 'react-spring'
 import { animated } from 'react-spring'
 import { BooleanControl } from '../../../controls/boolean-control'
 import { ColorControl } from '../../../controls/color-control'
-import type { ControlStatus, ControlStyles } from '../../../common/control-status'
+import type { ControlStatus } from '../../../common/control-status'
+import type { ControlStyles } from '../../../common/control-styles'
 import { PropertyRow, PropertyRowHeightWithLabel } from '../../../widgets/property-row'
 import { useArraySuperControl } from '../../../controls/array-supercontrol'
 import type {
