+++ editor/src/uuiui/inputs/string-input.tsx
@@ -4,7 +4,8 @@ import { jsx } from '@emotion/react'
 import styled from '@emotion/styled'
 import composeRefs from '@seznam/compose-react-refs'
 import React from 'react'
-import type { ControlStatus, ControlStyles } from '../../components/inspector/common/control-status'
+import type { ControlStatus } from '../../components/inspector/common/control-status'
+import type { ControlStyles } from '../../components/inspector/common/control-styles'
 import { getControlStyles } from '../../components/inspector/common/control-styles'
 import { preventDefault, stopPropagation } from '../../components/inspector/common/inspector-utils'
 import { useColorTheme, UtopiaTheme } from '../styles/theme'
